Cleanup the Gradle script, code style, project structure
