+++ .editorconfig
@@ -4,10 +4,10 @@ root = true
 end_of_line = lf
 insert_final_newline = true
 charset = utf-8
-indent_size = 4
+indent_size = 2
 indent_style = space
 trim_trailing_whitespace = true
-continuation_indent_size = 8
+continuation_indent_size = 4
 
 [*.yml]
 indent_size = 2
