+++ gradle.properties
@@ -1,4 +1,14 @@
 org.gradle.daemon=true
-org.gradle.jvmargs=-XX:MaxPermSize=256m -Xms512m -Xmx512m
-bintray.user=
-bintray.key=
+
+group=de.gesellix
+
+github.package-registry.owner=gesellix
+github.package-registry.repository=couchdb-client
+github.package-registry.username=
+github.package-registry.password=
+
+sonatype.snapshot.url=https://oss.sonatype.org/content/repositories/snapshots/
+sonatype.staging.url=https://oss.sonatype.org/service/local/staging/deploy/maven2/
+sonatype.staging.profile.id=
+sonatype.username=
+sonatype.password=
