+++ src/main/groovy/de/gesellix/couchdb/CouchDBClient.groovy
@@ -17,456 +17,456 @@ import static okhttp3.MediaType.parse
 @Slf4j
 class CouchDBClient {
 
-    OkHttpClient client
+  OkHttpClient client
 
-    Moshi moshi
+  Moshi moshi
 
-    String couchdbHost
-    int couchdbPort
+  String couchdbHost
+  int couchdbPort
 
-    String couchdbUsername
-    String couchdbPassword
+  String couchdbUsername
+  String couchdbPassword
 
-    CouchDBClient() {
-        this.client = new OkHttpClient()
-        this.moshi = new Moshi.Builder()
-                .build()
-        this.couchdbHost = "127.0.0.1"
-        this.couchdbPort = 5984
+  CouchDBClient() {
+    this.client = new OkHttpClient()
+    this.moshi = new Moshi.Builder()
+        .build()
+    this.couchdbHost = "127.0.0.1"
+    this.couchdbPort = 5984
+  }
+
+  def healthy() {
+    def builder = new Request.Builder()
+        .url("http://${couchdbHost}:${couchdbPort}")
+        .get()
+    def request = builder.build()
+
+    def response = client.newCall(request).execute()
+
+    if (!response.successful) {
+      if (response.code() == 404) {
+        response.body().close()
+        return false
+      }
+      else {
+        log.error("error {}", response.message())
+        response.body().close()
+        throw new IllegalStateException("could not check for database existence")
+      }
+    }
+    else {
+      log.info(response.body().string())
+      return true
+    }
+  }
+
+  def <R> R query(String db, String viewName, String key, boolean includeDocs = true) {
+    String encodedKey = moshi.adapter(String).toJson(key)
+
+    Request request = new Request.Builder()
+        .url("http://${couchdbHost}:${couchdbPort}" +
+             "/${db.toLowerCase()}" +
+             "/_design/${db.capitalize()}" +
+             "/_view/${viewName}" +
+             "?include_docs=${includeDocs}" +
+             "&key=${encodedKey}")
+        .build()
+
+    def response = client.newCall(request).execute()
+    if (!response.successful) {
+      log.error("request failed: {}", request)
+      throw new IOException(response.body().string())
     }
 
-    def healthy() {
-        def builder = new Request.Builder()
-                .url("http://${couchdbHost}:${couchdbPort}")
-                .get()
-        def request = builder.build()
+    def result = consume(response.body().byteStream())
+    result.rows.collect { row -> includeDocs ? row.doc : row }
+  }
 
-        def response = client.newCall(request).execute()
+  def <R> R query(String db, String viewName, Collection<String> keys, boolean includeDocs = true) {
+    String encodedKeys = moshi.adapter(Collection).toJson(keys)
 
-        if (!response.successful) {
-            if (response.code() == 404) {
-                response.body().close()
-                return false
-            }
-            else {
-                log.error("error {}", response.message())
-                response.body().close()
-                throw new IllegalStateException("could not check for database existence")
-            }
-        }
-        else {
-            log.info(response.body().string())
-            return true
-        }
+    Request request = new Request.Builder()
+        .url("http://${couchdbHost}:${couchdbPort}" +
+             "/${db.toLowerCase()}" +
+             "/_design/${db.capitalize()}" +
+             "/_view/${viewName}" +
+             "?include_docs=${includeDocs}" +
+             "&keys=${encodedKeys}")
+        .build()
+
+    def response = client.newCall(request).execute()
+    if (!response.successful) {
+      log.error("request failed: {}", request)
+      throw new IOException(response.body().string())
     }
 
-    def <R> R query(String db, String viewName, String key, boolean includeDocs = true) {
-        String encodedKey = moshi.adapter(String).toJson(key)
+    def result = consume(response.body().byteStream())
+    result.rows.collect { row -> includeDocs ? row.doc : row }
+  }
 
-        Request request = new Request.Builder()
-                .url("http://${couchdbHost}:${couchdbPort}" +
-                     "/${db.toLowerCase()}" +
-                     "/_design/${db.capitalize()}" +
-                     "/_view/${viewName}" +
-                     "?include_docs=${includeDocs}" +
-                     "&key=${encodedKey}")
-                .build()
+  def getAllDocs(String db, boolean includeDocs = true, boolean includeDesignDoc = false) {
+    def builder = new Request.Builder()
+        .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}" +
+             "/_all_docs" +
+             "?include_docs=${includeDocs}")
+        .get()
+    if (couchdbUsername && couchdbPassword) {
+      builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
+    }
+    def request = builder.build()
 
-        def response = client.newCall(request).execute()
-        if (!response.successful) {
-            log.error("request failed: {}", request)
-            throw new IOException(response.body().string())
-        }
+    def response = client.newCall(request).execute()
 
-        def result = consume(response.body().byteStream())
-        result.rows.collect { row -> includeDocs ? row.doc : row }
+    if (!response.successful) {
+      log.error("error getting all docs: {}/{}", response.code(), response.message())
+      throw new IllegalStateException("could not get all docs with")
+    }
+    else {
+      def allDocs = consume(response.body().byteStream())
+      allDocs = allDocs.rows.collect { row -> includeDocs ? row.doc : row }
+      if (!includeDesignDoc) {
+        allDocs = allDocs.grep { !it._id?.startsWith("_design/") && !it.id?.startsWith("_design/") }
+      }
+      return allDocs
+    }
+  }
+
+  def create(String db, def document) {
+    if (document == null) {
+      throw new IllegalArgumentException("document may not be null")
+    }
+    if (document['_rev']) {
+      log.error("document must be new, but has id({}), rev({})", document['_id'], document['_rev'])
+      throw new IllegalArgumentException("document must be new")
     }
 
-    def <R> R query(String db, String viewName, Collection<String> keys, boolean includeDocs = true) {
-        String encodedKeys = moshi.adapter(Collection).toJson(keys)
+    beforeCreate(document)
 
-        Request request = new Request.Builder()
-                .url("http://${couchdbHost}:${couchdbPort}" +
-                     "/${db.toLowerCase()}" +
-                     "/_design/${db.capitalize()}" +
-                     "/_view/${viewName}" +
-                     "?include_docs=${includeDocs}" +
-                     "&keys=${encodedKeys}")
-                .build()
+    def documentAsJson = moshi.adapter(Map).toJson(document)
+    def body = RequestBody.create(documentAsJson, parse("application/json"))
 
-        def response = client.newCall(request).execute()
-        if (!response.successful) {
-            log.error("request failed: {}", request)
-            throw new IOException(response.body().string())
-        }
+    def builder = new Request.Builder()
+    if (document['_id']) {
+      String docId = document['_id']
+      docId = sanitizeDocId(docId)
+      builder = builder
+          .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}/${docId}")
+          .put(body)
+    }
+    else {
+      builder = builder
+          .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}")
+          .post(body)
+    }
+    if (couchdbUsername && couchdbPassword) {
+      builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
+    }
+    def request = builder.build()
 
-        def result = consume(response.body().byteStream())
-        result.rows.collect { row -> includeDocs ? row.doc : row }
+    def response = client.newCall(request).execute()
+
+    def result = consume(response.body().byteStream())
+    if (!result.ok) {
+      log.error("error {}", result)
+      throw new IllegalStateException("error creating document")
     }
 
-    def getAllDocs(String db, boolean includeDocs = true, boolean includeDesignDoc = false) {
-        def builder = new Request.Builder()
-                .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}" +
-                     "/_all_docs" +
-                     "?include_docs=${includeDocs}")
-                .get()
-        if (couchdbUsername && couchdbPassword) {
-            builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
-        }
-        def request = builder.build()
+    def createdDocument = moshi.adapter(Object).fromJson(documentAsJson)
+    createdDocument['_id'] = result.id
+    createdDocument['_rev'] = result.rev
+    return createdDocument
+  }
 
-        def response = client.newCall(request).execute()
-
-        if (!response.successful) {
-            log.error("error getting all docs: {}/{}", response.code(), response.message())
-            throw new IllegalStateException("could not get all docs with")
-        }
-        else {
-            def allDocs = consume(response.body().byteStream())
-            allDocs = allDocs.rows.collect { row -> includeDocs ? row.doc : row }
-            if (!includeDesignDoc) {
-                allDocs = allDocs.grep { !it._id?.startsWith("_design/") && !it.id?.startsWith("_design/") }
-            }
-            return allDocs
-        }
+  def update(String db, def document) {
+    if (document == null) {
+      throw new IllegalArgumentException("document may not be null")
+    }
+    if (!document['_id']) {
+      throw new IllegalArgumentException("document id missing")
     }
 
-    def create(String db, def document) {
-        if (document == null) {
-            throw new IllegalArgumentException("document may not be null")
-        }
-        if (document['_rev']) {
-            log.error("document must be new, but has id({}), rev({})", document['_id'], document['_rev'])
-            throw new IllegalArgumentException("document must be new")
-        }
+    beforeUpdate(document)
 
-        beforeCreate(document)
+    def documentAsJson = moshi.adapter(Map).toJson(document)
+    def body = RequestBody.create(documentAsJson, parse("application/json"))
 
-        def documentAsJson = moshi.adapter(Map).toJson(document)
-        def body = RequestBody.create(documentAsJson, parse("application/json"))
+    def builder = new Request.Builder()
+    String docId = document['_id']
+    docId = sanitizeDocId(docId)
+    builder = builder
+        .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}/${docId}")
+        .put(body)
+    if (couchdbUsername && couchdbPassword) {
+      builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
+    }
+    def request = builder.build()
 
-        def builder = new Request.Builder()
-        if (document['_id']) {
-            String docId = document['_id']
-            docId = sanitizeDocId(docId)
-            builder = builder
-                    .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}/${docId}")
-                    .put(body)
-        }
-        else {
-            builder = builder
-                    .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}")
-                    .post(body)
-        }
-        if (couchdbUsername && couchdbPassword) {
-            builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
-        }
-        def request = builder.build()
+    def response = client.newCall(request).execute()
 
-        def response = client.newCall(request).execute()
-
-        def result = consume(response.body().byteStream())
-        if (!result.ok) {
-            log.error("error {}", result)
-            throw new IllegalStateException("error creating document")
-        }
-
-        def createdDocument = moshi.adapter(Object).fromJson(documentAsJson)
-        createdDocument['_id'] = result.id
-        createdDocument['_rev'] = result.rev
-        return createdDocument
+    def result = consume(response.body().byteStream())
+    if (!result.ok) {
+      log.error("error {}", result)
+      throw new IllegalStateException("error updating document")
     }
 
-    def update(String db, def document) {
-        if (document == null) {
-            throw new IllegalArgumentException("document may not be null")
-        }
-        if (!document['_id']) {
-            throw new IllegalArgumentException("document id missing")
-        }
+    def updatedDocument = moshi.adapter(Object).fromJson(documentAsJson)
+    updatedDocument['_rev'] = result.rev
+    return updatedDocument
+  }
 
+  def updateBulk(String db, List documents) {
+    if (documents == null) {
+      throw new IllegalArgumentException("documents may not be null")
+    }
+    if (documents.empty) {
+      throw new IllegalArgumentException("documents is empty")
+    }
+
+    documents.each { document ->
+      if (document['_id']) {
         beforeUpdate(document)
-
-        def documentAsJson = moshi.adapter(Map).toJson(document)
-        def body = RequestBody.create(documentAsJson, parse("application/json"))
-
-        def builder = new Request.Builder()
-        String docId = document['_id']
-        docId = sanitizeDocId(docId)
-        builder = builder
-                .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}/${docId}")
-                .put(body)
-        if (couchdbUsername && couchdbPassword) {
-            builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
-        }
-        def request = builder.build()
-
-        def response = client.newCall(request).execute()
-
-        def result = consume(response.body().byteStream())
-        if (!result.ok) {
-            log.error("error {}", result)
-            throw new IllegalStateException("error updating document")
-        }
-
-        def updatedDocument = moshi.adapter(Object).fromJson(documentAsJson)
-        updatedDocument['_rev'] = result.rev
-        return updatedDocument
+      }
+      else if (!document.dateCreated) {
+        beforeCreate(document)
+      }
     }
 
-    def updateBulk(String db, List documents) {
-        if (documents == null) {
-            throw new IllegalArgumentException("documents may not be null")
-        }
-        if (documents.empty) {
-            throw new IllegalArgumentException("documents is empty")
-        }
+    def updateDoc = [docs: documents]
 
-        documents.each { document ->
-            if (document['_id']) {
-                beforeUpdate(document)
-            }
-            else if (!document.dateCreated) {
-                beforeCreate(document)
-            }
-        }
+    def documentAsJson = moshi.adapter(Object).toJson(updateDoc)
+    def body = RequestBody.create(documentAsJson, parse("application/json"))
 
-        def updateDoc = [docs: documents]
+    def builder = new Request.Builder()
+    builder = builder
+        .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}/_bulk_docs")
+        .post(body)
+    if (couchdbUsername && couchdbPassword) {
+      builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
+    }
+    def request = builder.build()
 
-        def documentAsJson = moshi.adapter(Object).toJson(updateDoc)
-        def body = RequestBody.create(documentAsJson, parse("application/json"))
-
-        def builder = new Request.Builder()
-        builder = builder
-                .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}/_bulk_docs")
-                .post(body)
-        if (couchdbUsername && couchdbPassword) {
-            builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
-        }
-        def request = builder.build()
-
-        def response = client.newCall(request).execute()
-        if (!response.successful) {
-            log.error("error {}", response.message())
-            throw new IllegalStateException("bulk update failed")
-        }
-
-        def result = consume(response.body().byteStream())
-        [result, documents].transpose().each { updated, original ->
-            if (updated.ok) {
-                original._id = updated.id
-                original._rev = updated.rev
-                return original
-            }
-            else {
-                log.error("error {}", updated)
-                return null
-            }
-        }
-
-        return result
+    def response = client.newCall(request).execute()
+    if (!response.successful) {
+      log.error("error {}", response.message())
+      throw new IllegalStateException("bulk update failed")
     }
 
-    def containsDb(String db) {
-        def builder = new Request.Builder()
-                .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}")
-                .head()
-        if (couchdbUsername && couchdbPassword) {
-            builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
-        }
-        def request = builder.build()
-
-        def response = client.newCall(request).execute()
-
-        if (!response.successful) {
-            if (response.code() == 404) {
-                return false
-            }
-            else {
-                log.error("error {}", response.message())
-                throw new IllegalStateException("could not check for database existence")
-            }
-        }
-        else {
-            return true
-        }
+    def result = consume(response.body().byteStream())
+    [result, documents].transpose().each { updated, original ->
+      if (updated.ok) {
+        original._id = updated.id
+        original._rev = updated.rev
+        return original
+      }
+      else {
+        log.error("error {}", updated)
+        return null
+      }
     }
 
-    def createDb(String db) {
-        def builder = new Request.Builder()
-                .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}")
-                .put(RequestBody.create('', parse("application/json")))
-        if (couchdbUsername && couchdbPassword) {
-            builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
-        }
-        def request = builder.build()
+    return result
+  }
 
-        def response = client.newCall(request).execute()
+  def containsDb(String db) {
+    def builder = new Request.Builder()
+        .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}")
+        .head()
+    if (couchdbUsername && couchdbPassword) {
+      builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
+    }
+    def request = builder.build()
 
-        def result = consume(response.body().byteStream())
-        if (!result.ok) {
-            log.error("error {}", result)
-            throw new IllegalStateException("error creating database")
-        }
-        return result
+    def response = client.newCall(request).execute()
+
+    if (!response.successful) {
+      if (response.code() == 404) {
+        return false
+      }
+      else {
+        log.error("error {}", response.message())
+        throw new IllegalStateException("could not check for database existence")
+      }
+    }
+    else {
+      return true
+    }
+  }
+
+  def createDb(String db) {
+    def builder = new Request.Builder()
+        .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}")
+        .put(RequestBody.create('', parse("application/json")))
+    if (couchdbUsername && couchdbPassword) {
+      builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
+    }
+    def request = builder.build()
+
+    def response = client.newCall(request).execute()
+
+    def result = consume(response.body().byteStream())
+    if (!result.ok) {
+      log.error("error {}", result)
+      throw new IllegalStateException("error creating database")
+    }
+    return result
+  }
+
+  def deleteDb(String db) {
+    def builder = new Request.Builder()
+        .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}")
+        .delete()
+    if (couchdbUsername && couchdbPassword) {
+      builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
+    }
+    def request = builder.build()
+
+    def response = client.newCall(request).execute()
+
+    def result = consume(response.body().byteStream())
+    if (!result.ok) {
+      log.error("error {}", result)
+      throw new IllegalStateException("error deleting database")
+    }
+    return result
+  }
+
+  def createFindAllDocsView(String db) {
+    def findAll = "function(doc) { emit(null, doc._id) }"
+    return createView(db, "all", findAll, null)
+  }
+
+  def createFindByPropertyView(String db, String propertyName) {
+    String findByProperty = "function(doc) { if (doc['${propertyName}']) { emit(doc['${propertyName}'], doc._id) } }"
+    return createView(db, "by_${propertyName}", findByProperty, null)
+  }
+
+  def createView(String db, String viewName, String mapFunction, String reduceFunction) {
+    def view = [:]
+    if (mapFunction) {
+      view['map'] = mapFunction
+    }
+    if (reduceFunction) {
+      view['reduce'] = reduceFunction
     }
 
-    def deleteDb(String db) {
-        def builder = new Request.Builder()
-                .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}")
-                .delete()
-        if (couchdbUsername && couchdbPassword) {
-            builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
-        }
-        def request = builder.build()
-
-        def response = client.newCall(request).execute()
-
-        def result = consume(response.body().byteStream())
-        if (!result.ok) {
-            log.error("error {}", result)
-            throw new IllegalStateException("error deleting database")
-        }
-        return result
-    }
-
-    def createFindAllDocsView(String db) {
-        def findAll = "function(doc) { emit(null, doc._id) }"
-        return createView(db, "all", findAll, null)
-    }
-
-    def createFindByPropertyView(String db, String propertyName) {
-        String findByProperty = "function(doc) { if (doc['${propertyName}']) { emit(doc['${propertyName}'], doc._id) } }"
-        return createView(db, "by_${propertyName}", findByProperty, null)
-    }
-
-    def createView(String db, String viewName, String mapFunction, String reduceFunction) {
-        def view = [:]
-        if (mapFunction) {
-            view['map'] = mapFunction
-        }
-        if (reduceFunction) {
-            view['reduce'] = reduceFunction
-        }
-
-        String designDocId = "_design/${db.capitalize()}"
-        def newDesignDoc = [
-                _id     : designDocId,
-                language: "javascript",
-                views   : [
-                        (viewName): view
-                ]
+    String designDocId = "_design/${db.capitalize()}"
+    def newDesignDoc = [
+        _id     : designDocId,
+        language: "javascript",
+        views   : [
+            (viewName): view
         ]
+    ]
 
-        def designDocExists = contains(db, designDocId)
-        if (designDocExists) {
-            def currentDesignDoc = get(db, designDocId)
-            def mergedDesignDoc = merge(currentDesignDoc, newDesignDoc)
-            if (mergedDesignDoc != currentDesignDoc) {
-                update(db, mergedDesignDoc)
-            }
-            return currentDesignDoc
-        }
-        else {
-            return update(db, newDesignDoc)
-        }
+    def designDocExists = contains(db, designDocId)
+    if (designDocExists) {
+      def currentDesignDoc = get(db, designDocId)
+      def mergedDesignDoc = merge(currentDesignDoc, newDesignDoc)
+      if (mergedDesignDoc != currentDesignDoc) {
+        update(db, mergedDesignDoc)
+      }
+      return currentDesignDoc
     }
-
-    def contains(String db, String docId) {
-        docId = sanitizeDocId(docId)
-        def builder = new Request.Builder()
-                .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}/${docId}")
-                .head()
-        if (couchdbUsername && couchdbPassword) {
-            builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
-        }
-        def request = builder.build()
-
-        def response = client.newCall(request).execute()
-
-        if (!response.successful) {
-            if (response.code() == 404) {
-                return false
-            }
-            else {
-                log.error("error {}", response.message())
-                throw new IllegalStateException("could not check for doc existence")
-            }
-        }
-        else {
-            return true
-        }
+    else {
+      return update(db, newDesignDoc)
     }
+  }
 
-    def get(String db, String docId) {
-        docId = sanitizeDocId(docId)
-        def builder = new Request.Builder()
-                .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}/${docId}")
-                .get()
-        if (couchdbUsername && couchdbPassword) {
-            builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
-        }
-        def request = builder.build()
-
-        def response = client.newCall(request).execute()
-
-        if (!response.successful) {
-            log.error("error getting document({}): {}/{}", docId, response.code(), response.message())
-            throw new IllegalStateException("could not get doc with id '${docId}'")
-        }
-        else {
-            return consume(response.body().byteStream())
-        }
+  def contains(String db, String docId) {
+    docId = sanitizeDocId(docId)
+    def builder = new Request.Builder()
+        .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}/${docId}")
+        .head()
+    if (couchdbUsername && couchdbPassword) {
+      builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
     }
+    def request = builder.build()
 
-    def delete(String db, String docId, String rev) {
-        docId = sanitizeDocId(docId)
-        def builder = new Request.Builder()
-                .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}/${docId}?rev=${rev}")
-                .delete()
-        if (couchdbUsername && couchdbPassword) {
-            builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
-        }
-        def request = builder.build()
+    def response = client.newCall(request).execute()
 
-        def response = client.newCall(request).execute()
-
-        if (!response.successful) {
-            log.error("error deleting doc{ _id:{}, _rev:{} }: {}/{}", docId, rev, response.code(), response.message())
-            throw new IllegalStateException("could not delete doc with id/rev '${docId}'/'${rev}'")
-        }
-        else {
-            def result = consume(response.body().byteStream())
-            return result
-        }
+    if (!response.successful) {
+      if (response.code() == 404) {
+        return false
+      }
+      else {
+        log.error("error {}", response.message())
+        throw new IllegalStateException("could not check for doc existence")
+      }
     }
-
-    static def merge(def currentDoc, def changedDoc) {
-        def mergedDoc = currentDoc + [:]
-        mergedDoc.views = (currentDoc.views ?: [:]) + (changedDoc.views ?: [:])
-        return mergedDoc
+    else {
+      return true
     }
+  }
 
-    def consume(InputStream stream) {
-        def result = moshi.adapter(Object).fromJson(Okio.buffer(Okio.source(stream)))
-        Util.closeQuietly(stream)
-        return result
+  def get(String db, String docId) {
+    docId = sanitizeDocId(docId)
+    def builder = new Request.Builder()
+        .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}/${docId}")
+        .get()
+    if (couchdbUsername && couchdbPassword) {
+      builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
     }
+    def request = builder.build()
 
-    static def sanitizeDocId(String docId) {
-        if (!docId.startsWith('_')) {
-            docId = URLEncoder.encode(docId, UTF_8.toString())
-        }
-        docId
-    }
+    def response = client.newCall(request).execute()
 
-    def beforeCreate(def document) {
-        if (!document.dateCreated) {
-            document.dateCreated = LocalDate.now().toString()
-        }
+    if (!response.successful) {
+      log.error("error getting document({}): {}/{}", docId, response.code(), response.message())
+      throw new IllegalStateException("could not get doc with id '${docId}'")
     }
+    else {
+      return consume(response.body().byteStream())
+    }
+  }
 
-    def beforeUpdate(def document) {
-        document.dateUpdated = LocalDate.now().toString()
+  def delete(String db, String docId, String rev) {
+    docId = sanitizeDocId(docId)
+    def builder = new Request.Builder()
+        .url("http://${couchdbHost}:${couchdbPort}/${db.toLowerCase()}/${docId}?rev=${rev}")
+        .delete()
+    if (couchdbUsername && couchdbPassword) {
+      builder = builder.header("Authorization", Credentials.basic(couchdbUsername, couchdbPassword))
     }
+    def request = builder.build()
+
+    def response = client.newCall(request).execute()
+
+    if (!response.successful) {
+      log.error("error deleting doc{ _id:{}, _rev:{} }: {}/{}", docId, rev, response.code(), response.message())
+      throw new IllegalStateException("could not delete doc with id/rev '${docId}'/'${rev}'")
+    }
+    else {
+      def result = consume(response.body().byteStream())
+      return result
+    }
+  }
+
+  static def merge(def currentDoc, def changedDoc) {
+    def mergedDoc = currentDoc + [:]
+    mergedDoc.views = (currentDoc.views ?: [:]) + (changedDoc.views ?: [:])
+    return mergedDoc
+  }
+
+  def consume(InputStream stream) {
+    def result = moshi.adapter(Object).fromJson(Okio.buffer(Okio.source(stream)))
+    Util.closeQuietly(stream)
+    return result
+  }
+
+  static def sanitizeDocId(String docId) {
+    if (!docId.startsWith('_')) {
+      docId = URLEncoder.encode(docId, UTF_8.toString())
+    }
+    docId
+  }
+
+  def beforeCreate(def document) {
+    if (!document.dateCreated) {
+      document.dateCreated = LocalDate.now().toString()
+    }
+  }
+
+  def beforeUpdate(def document) {
+    document.dateUpdated = LocalDate.now().toString()
+  }
 }
